# src/main/java/com/kshitij/studentmanagement/model/Student.java
package com.kshitij.studentmanagement.model;

import jakarta.persistence.*;
import lombok.Data;

@Data
@Entity
public class Student {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(nullable = false)
    private String name;

    @Column(nullable = false)
    private int age;

    @Column(nullable = false)
    private String className;

    @Column(nullable = false)
    private String phoneNumber;
}